VariantImporterTest created (moved from main)
